requirements.txt
dj-database-url==0.5.0
django-allauth==0.40.0
Django==2.2.7
gunicorn==19.9.0
whitenoise==4.1.4
Pillow==6.2.1
psycopg2